refactor(tui): rename `tui/tuiutil` to `tui/util`
