+++ tui/util/keymap.go
@@ -1,4 +1,4 @@
-package tuiutil
+package util
 
 import "github.com/charmbracelet/bubbles/key"
